requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
psycopg[binary]==3.2.1
sqlglot==25.10.0
openai==1.45.0
python-dotenv==1.0.1
httpx<0.28
